+++ src/Parser/ClassParser.php
@@ -102,10 +102,14 @@ class ClassParser extends AbstractStructureParser
 
         // So we got our docBlock, now we can parse the invariant annotations from it
         $annotationParser = new AnnotationParser($this->file, $this->config, $this->tokens, $this->currentDefinition);
-        $this->currentDefinition->invariantConditions = $annotationParser->getConditions(
+        $invariantConditions = $annotationParser->getConditions(
             $this->currentDefinition->getDocBlock(),
             Annotations::INVARIANT
         );
+        if (!is_bool($invariantConditions)) {
+
+            $this->currentDefinition->invariantConditions = $invariantConditions;
+        }
 
         // we would be also interested in introductions
         $introductions = new IntroductionList();